backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
bcrypt==4.2.0
PyJWT==2.9.0
SQLAlchemy==2.0.36
